Added internet perms, material progress bar for laod indication, recycler adapter, updated strings, utilty for rounding values.

+++ app/src/main/java/lokovsek/pir/loadingpivo/api/dto/Rating.kt
@@ -11,7 +11,7 @@ data class Rating(
 
     @Expose
     @SerializedName("reviews")
-    val reviews: Long? = null,
+    val reviews: Double? = null,
 
     ) {
     override fun toString(): String {
